prajwal/policy-engine/requirements.txt
"""Requirements for MLOps Policy Engine."""
pyyaml>=5.4
flask>=2.0.0
click>=8.0.0
tabulate>=0.8.9
pytest>=6.2.0
requests>=2.25.0
